# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/user/create": {
            "post": {
                "description": "Add a new user item to the list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User item to create",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dts.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created user item",
                        "schema": {
                            "$ref": "#/definitions/dts.User"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/user/get": {
            "get": {
                "description": "Retrieve a list of all user items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dts.User"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dts.User": {
            "type": "object",
            "required": [
                "age",
                "firstname",
                "lastname"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 24
                },
                "firstname": {
                    "type": "string",
                    "example": "This is the user firstname"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "lastname": {
                    "type": "string",
                    "example": "This is the user lastname"
                }
            }
        }
    }
}